restructured splash page into different files
