+++ frontend/tests.js
@@ -11,6 +11,8 @@ import { ArtistHome, ArtistItem } from './src/components/home-pages/ArtistsHome'
 import { MediaHome, MediaItem } from './src/components/home-pages/MediaHome'
 import { UIRouter } from '@uirouter/react';
 import Splash from './src/components/Splash';
+import SplashCarousel from './src/components/SplashCarousel';
+import SplashDescription from './src/components/SplashDescription';
 import About from './src/components/About';
 import { Desc, Bio, Statistics, Data, Tools } from './src/components/About';
 import { MediaInstance } from './src/components/instance-pages/MediaInstance';
@@ -160,12 +162,17 @@ describe('<MediaHome/>', function () {
 
 // Splash
 describe('<Splash/>', function () {
+  it('should render without crashing', function () {
+    shallow(<Splash />);
+  })
+
+  describe('<SplashCarousel/>', function () {
     it('should render without crashing', function () {
-        shallow(<Splash />);
+      shallow(<SplashCarousel />);
     })
 
     it('should have 4 items: Albums, Artists, Movies and TV, and Making Connections', function () {
-      const wrapper = shallow(<Splash />);
+      const wrapper = shallow(<SplashCarousel />);
       const carouselItems = wrapper.find('CarouselCaption');
       expect(carouselItems).to.have.length(4);
       expect(carouselItems.at(0).render().text()).to.equal('Albums');
@@ -174,55 +181,62 @@ describe('<Splash/>', function () {
       expect(carouselItems.at(3).render().text()).to.equal('Making Connections');
     })
 
-  describe('<Carousel/>', () => {
-    const items = [
-      { src: '', altText: 'item A', caption: 'caption A' },
-      { src: '', altText: 'item B', caption: 'caption B' },
-      { src: '', altText: 'item C', caption: 'caption C' }
-    ];
+    describe('<Carousel/>', () => {
+      const items = [
+        { src: '', altText: 'item A', caption: 'caption A' },
+        { src: '', altText: 'item B', caption: 'caption B' },
+        { src: '', altText: 'item C', caption: 'caption C' }
+      ];
 
-    it('should render a header and a caption', () => {
-      const wrapper = mount(<CarouselCaption captionHeader="headerText" captionText="bodyText" />);
-      expect(wrapper.find('h3').length).to.equal(1);
-      expect(wrapper.find('p').length).to.equal(1);
-    });
-
-    it('should render a list with the right number of items', () => {
-      const wrapper = mount(<CarouselIndicators items={items} activeIndex={0} onClickHandler={() => { }} />);
-      expect(wrapper.find('ol').length).to.equal(1);
-      expect(wrapper.find('li').length).to.equal(3);
-    });
-
-    it('should append the correct active class', () => {
-      const wrapper = mount(<CarouselIndicators items={items} activeIndex={0} onClickHandler={() => { }} />);
-      expect(wrapper.find('.active').hostNodes().length).to.equal(1);
-    });
-
-    it('should render an anchor tag', () => {
-      const wrapper = mount(<CarouselControl direction="next" onClickHandler={() => { }} />);
-      expect(wrapper.find('a').length).to.equal(1);
-    });
-
-    it('should call the onClickHandler', () => {
-      const onClick = sinon.spy();
-      const slides = items.map((item, idx) => {
-        return (
-          <CarouselItem onExiting={() => { }} onExited={() => { }} key={idx}>
-            <CarouselCaption captionText={item.caption} captionHeader={item.caption} />
-          </CarouselItem>
-        );
+      it('should render a header and a caption', () => {
+        const wrapper = mount(<CarouselCaption captionHeader="headerText" captionText="bodyText" />);
+        expect(wrapper.find('h3').length).to.equal(1);
+        expect(wrapper.find('p').length).to.equal(1);
       });
 
-      const wrapper = mount(
-        <Carousel activeIndex={0} next={() => { }} previous={() => { }}>
-          <CarouselIndicators items={items} activeIndex={0} onClickHandler={() => { }} />
-          {slides}
-          <CarouselControl direction="next" directionText="Next" onClickHandler={onClick} />
-        </Carousel>
-      );
-      wrapper.find('a').first().simulate('click');
-      sinon.assert.called(onClick);
+      it('should render a list with the right number of items', () => {
+        const wrapper = mount(<CarouselIndicators items={items} activeIndex={0} onClickHandler={() => { }} />);
+        expect(wrapper.find('ol').length).to.equal(1);
+        expect(wrapper.find('li').length).to.equal(3);
+      });
+
+      it('should append the correct active class', () => {
+        const wrapper = mount(<CarouselIndicators items={items} activeIndex={0} onClickHandler={() => { }} />);
+        expect(wrapper.find('.active').hostNodes().length).to.equal(1);
+      });
+
+      it('should render an anchor tag', () => {
+        const wrapper = mount(<CarouselControl direction="next" onClickHandler={() => { }} />);
+        expect(wrapper.find('a').length).to.equal(1);
+      });
+
+      it('should call the onClickHandler', () => {
+        const onClick = sinon.spy();
+        const slides = items.map((item, idx) => {
+          return (
+            <CarouselItem onExiting={() => { }} onExited={() => { }} key={idx}>
+              <CarouselCaption captionText={item.caption} captionHeader={item.caption} />
+            </CarouselItem>
+          );
+        });
+
+        const wrapper = mount(
+          <Carousel activeIndex={0} next={() => { }} previous={() => { }}>
+            <CarouselIndicators items={items} activeIndex={0} onClickHandler={() => { }} />
+            {slides}
+            <CarouselControl direction="next" directionText="Next" onClickHandler={onClick} />
+          </Carousel>
+        );
+        wrapper.find('a').first().simulate('click');
+        sinon.assert.called(onClick);
+      });
     });
+  });
+
+  describe('<SplashDescription/>', function () {
+    it('should render without crashing', function () {
+        shallow(<SplashDescription />);
+    })
   });
 });
